implementation of telegraf library NOT WORKING

+++ pkg/collector/processes/telegraf/types.go
@@ -5,7 +5,7 @@ type ProcessInfo struct {
 	PID         int32   `json:"pid"`
 	User        string  `json:"user"`
 	CPUUsage    float64 `json:"cpuUsage"`
-	MemoryUsage float32 `json:"memUsage"`
+	MemoryUsage float64 `json:"memUsage"`
 	Command     string  `json:"command"`
 	Status      string  `json:"status"`
 }
